refactor(config): split grid.py into sweep-expansion and cycle-ordering
Decompose the 786-line config/grid.py god-module along its natural seams
into three single-concern modules. Pure decomposition: no behavior or
signature change, study_design_hash values byte-identical.

- grid.py: grid-orchestration facade (expand_grid, SkippedConfig + rule-id
  classification, compute_study_design_hash, base/fixed helpers); re-exports
  the moved public API so the config.grid import surface is unchanged.
- config/sweep_expansion.py: sweep-block expansion into raw config dicts
  (axes, groups, key routing) plus count_sweep_structure.
- config/cycle_ordering.py: execution-sequence ordering across cycles
  (ExperimentOrder, apply_cycles, cycle_boundary_indices, Williams square).

Sweep-internal tests repoint their private-helper imports to the new home;
public import paths are untouched.
